chore(household): consolidate components
