+++ lib/household/bloc/household_state.dart
